producer/requirements.txt
kafka-python
dataprep
requests